% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m3_addInfo.R
\name{addInfo_3}
\alias{addInfo_3}
\title{add functions, schema and tables
create 13 funtions, shema mimic3 and 26 tables}
\usage{
addInfo_3(user = "postgres", password = "pg")
}
\arguments{
\item{conn}{connection, if missing, it will be get 'connectiontopsql' from global environment.}
}
\value{
13 funtions, shema mimic3 and 26 tables in PostgreSQL.
}
\description{
add functions, schema and tables
create 13 funtions, shema mimic3 and 26 tables
}
\examples{
\donttest{
addInfo_3()
}
}
