% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_addInfo.R
\name{addInfo_4}
\alias{addInfo_4}
\title{add functions, schema and tables
create shema mimic4 and 27 tables}
\usage{
addInfo_4(user = "postgres", password = "pg")
}
\arguments{
\item{conn}{connection, if missing, it will be get 'connect_MIMIC' from global environment.}
}
\value{
shema mimic4 and 27 tables in PostgreSQL.
}
\description{
add functions, schema and tables
create shema mimic4 and 27 tables
}
\examples{
\donttest{
addInfo_4()
}
}
