% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_addMatView.R
\name{addMatView_4}
\alias{addMatView_4}
\title{add materialized view data for mimic IV}
\usage{
addMatView_4(range, conn)
}
\arguments{
\item{range}{which to be added}

\item{conn}{connection of MIMIC, if missing, it well be got from global environment by connect_MIMIC}
}
\value{
58 materialized view data
}
\description{
add materialized view data for mimic IV
}
\examples{
\donttest{
addMatView_4()
}
}
