% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_MIMIC.R
\name{connect_MIMIC}
\alias{connect_MIMIC}
\title{Connect to MIMIC3 database in PostgreSQL}
\usage{
connect_MIMIC(user = "postgres", password = "pg", dbname = "mimic", ...)
}
\arguments{
\item{user}{user name}

\item{password}{password}

\item{dbname}{database name, default is mimic3.}

\item{...}{other parameter.}
}
\value{
connection to MIMIC3 database in PostgreSQL.
}
\description{
Connect to MIMIC3 database in PostgreSQL
}
\examples{
\donttest{
connect_MIMIC <- connect3()
}
}
