% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m3_delete_matview.R
\name{delete_matview_3}
\alias{delete_matview_3}
\title{delete materialized view from mimic3}
\usage{
delete_matview_3(matview, conn)
}
\arguments{
\item{matview}{matview names}

\item{conn}{connection}
}
\description{
delete materialized view from mimic3
}
\examples{
\donttest{
delete_matview_3()
}
}
