% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_MIMIC.R
\name{delete_MIMIC}
\alias{delete_MIMIC}
\title{Drop mimic database}
\usage{
delete_MIMIC(dbname = "mimic", user = "postgres", password = "pg")
}
\arguments{
\item{dbname}{mimic}

\item{user}{postgres}

\item{password}{pg}
}
\value{
drop database in PostgreSQL
}
\description{
Drop mimic database
}
\examples{
\donttest{
delete_MIMIC()
}
}
