% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_Dim.R
\name{Dim}
\alias{Dim}
\title{dimension of tbl_PostgreSQLConnection}
\usage{
Dim(x)
}
\arguments{
\item{x}{tbl_PostgreSQLConnection}
}
\value{
number of row and column
}
\description{
dimension of tbl_PostgreSQLConnection
}
