% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_depth}
\alias{get_depth}
\title{determine the depth(distance to root) of a term in the snomed hierarchy}
\usage{
get_depth(relationship_file, terms)
}
\description{
determine the depth(distance to root) of a term in the snomed hierarchy
}
