% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_freetext_diagnosis_as_snomed_ids_filter_and_write}
\alias{get_freetext_diagnosis_as_snomed_ids_filter_and_write}
\title{Parse note-events files, extract discharge diagnosis free text, extract UMLS CUIs with metamap, map cuis to SNOMED}
\usage{
get_freetext_diagnosis_as_snomed_ids_filter_and_write(
  noteevents_file,
  record_tags,
  umlsCUI_snomedID_mapfile,
  keep_types_list,
  outfile,
  add_tags_to_bd
)
}
\description{
Parse note-events files, extract discharge diagnosis free text, extract UMLS CUIs with metamap, map cuis to SNOMED
}
