% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m3_show_matview.R
\name{matview_3}
\alias{matview_3}
\title{Return materialized view names}
\usage{
matview_3(conn, schema = "mimic3")
}
\arguments{
\item{conn}{connection of mimic3 database}

\item{schema}{schema name, default is mimic3}
}
\value{
characters
}
\description{
Return materialized view names
}
