% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_Order.R
\name{Order}
\alias{Order}
\title{order for PostgreSQL data
order for data with tbl_PostgreSQLConnection}
\usage{
Order(x, ...)
}
\arguments{
\item{x}{data with tbl_PostgreSQLConnection}

\item{...}{one or more colnames start with minus or not}
}
\value{
data with tbl_PostgreSQLConnection
}
\description{
order for PostgreSQL data
order for data with tbl_PostgreSQLConnection
}
\examples{
\donttest{
library(mimicR)
connect_MIMIC <- connect_MIMIC()
tbl3_caregivers() \%>\%
    Order('label','-cgid')
}
}
