% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{parse_mimic_noteevents}
\alias{parse_mimic_noteevents}
\title{Return string \code{"Hello world!"} to R.}
\usage{
parse_mimic_noteevents(file, tags, keep_rawtext, outfile, add_tags_to_bd)
}
\description{
Return string \code{"Hello world!"} to R.
}
