% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_SeqN.R
\name{SeqN}
\alias{SeqN}
\title{Title}
\usage{
SeqN(x, column, newname = "SeqN")
}
\arguments{
\item{x}{dataframe or tbl_PostgreSQLConnection object}

\item{column}{one column name}

\item{newname}{new column name}
}
\value{
dataframe or tbl_PostgreSQLConnection with seqence
}
\description{
Title
}
