% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_postgre_q.R
\name{sql_postgre_q}
\alias{sql_postgre_q}
\title{Return q number of sql_render()}
\usage{
sql_postgre_q(x)
}
\arguments{
\item{x}{sql object}
}
\value{
one number
}
\description{
Return q number of sql_render()
}
