% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_to_command.R
\name{sql_to_command}
\alias{sql_to_command}
\title{put sql file lines to one command for r
sql files must be simplified.}
\usage{
sql_to_command(sql)
}
\arguments{
\item{sql}{sql file names}
}
\description{
put sql file lines to one command for r
sql files must be simplified.
}
