% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_frame.R
\name{[.tbl_PostgreSQLConnection}
\alias{[.tbl_PostgreSQLConnection}
\title{Extract Parts of tbl_PostgreSQLConnection}
\usage{
\method{[}{tbl_PostgreSQLConnection}(x, i, j = "*", view = FALSE, df = FALSE, head)
}
\arguments{
\item{x}{an object of tbl_PostgreSQLConnection}

\item{i}{row select}

\item{j}{column names}

\item{view}{logical, if TRUE, view data}

\item{df}{logical, if TRUE, convert the results to dataframe}

\item{head}{one number to extract the head data}
}
\value{
still tbl_PostgreSQLConnection object
}
\description{
Extract Parts of tbl_PostgreSQLConnection
}
