% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m3_tbl.R, R/m4_tbl.R
\name{table}
\alias{table}
\alias{tbl3_admissions}
\alias{tbl3_callout}
\alias{tbl3_caregivers}
\alias{tbl3_chartevents}
\alias{tbl3_cptevents}
\alias{tbl3_datetimeevents}
\alias{tbl3_diagnoses_icd}
\alias{tbl3_drgcodes}
\alias{tbl3_d_cpt}
\alias{tbl3_d_icd_diagnoses}
\alias{tbl3_procedures_icd}
\alias{tbl3_d_icd_procedures}
\alias{tbl3_d_items}
\alias{tbl3_d_labitems}
\alias{tbl3_icustays}
\alias{tbl3_inputevents_cv}
\alias{tbl3_inputevents_mv}
\alias{tbl3_labevents}
\alias{tbl3_microbiologyevents}
\alias{tbl3_noteevents}
\alias{tbl3_outputevents}
\alias{tbl3_patients}
\alias{tbl3_prescriptions}
\alias{tbl3_procedureevents_mv}
\alias{tbl3_services}
\alias{tbl3_transfers}
\alias{tbl4_admissions}
\alias{tbl4_callout}
\alias{tbl4_caregivers}
\alias{tbl4_chartevents}
\alias{tbl4_cptevents}
\alias{tbl4_datetimeevents}
\alias{tbl4_diagnoses_icd}
\alias{tbl4_drgcodes}
\alias{tbl4_d_cpt}
\alias{tbl4_d_icd_diagnoses}
\alias{tbl4_procedures_icd}
\alias{tbl4_d_icd_procedures}
\alias{tbl4_d_items}
\alias{tbl4_d_labitems}
\alias{tbl4_icustays}
\alias{tbl4_inputevents_cv}
\alias{tbl4_inputevents_mv}
\alias{tbl4_labevents}
\alias{tbl4_microbiologyevents}
\alias{tbl4_noteevents}
\alias{tbl4_outputevents}
\alias{tbl4_patients}
\alias{tbl4_prescriptions}
\alias{tbl4_procedureevents_mv}
\alias{tbl4_services}
\alias{tbl4_transfers}
\title{conn tables}
\usage{
tbl3_admissions(..., conn)

tbl3_callout(..., conn)

tbl3_caregivers(..., conn)

tbl3_chartevents(..., conn)

tbl3_cptevents(..., conn)

tbl3_datetimeevents(..., conn)

tbl3_diagnoses_icd(..., conn)

tbl3_drgcodes(..., conn)

tbl3_d_cpt(..., conn)

tbl3_d_icd_diagnoses(..., conn)

tbl3_procedures_icd(..., conn)

tbl3_d_icd_procedures(..., conn)

tbl3_d_items(..., conn)

tbl3_d_labitems(..., conn)

tbl3_icustays(..., conn)

tbl3_inputevents_cv(..., conn)

tbl3_inputevents_mv(..., conn)

tbl3_labevents(..., conn)

tbl3_microbiologyevents(..., conn)

tbl3_noteevents(..., conn)

tbl3_outputevents(..., conn)

tbl3_patients(..., conn)

tbl3_prescriptions(..., conn)

tbl3_procedureevents_mv(..., conn)

tbl3_services(..., conn)

tbl3_transfers(..., conn)

tbl4_admissions(..., conn)

tbl4_callout(..., conn)

tbl4_caregivers(..., conn)

tbl4_chartevents(..., conn)

tbl4_cptevents(..., conn)

tbl4_datetimeevents(..., conn)

tbl4_diagnoses_icd(..., conn)

tbl4_drgcodes(..., conn)

tbl4_d_cpt(..., conn)

tbl4_d_icd_diagnoses(..., conn)

tbl4_procedures_icd(..., conn)

tbl4_d_icd_procedures(..., conn)

tbl4_d_items(..., conn)

tbl4_d_labitems(..., conn)

tbl4_icustays(..., conn)

tbl4_inputevents_cv(..., conn)

tbl4_inputevents_mv(..., conn)

tbl4_labevents(..., conn)

tbl4_microbiologyevents(..., conn)

tbl4_noteevents(..., conn)

tbl4_outputevents(..., conn)

tbl4_patients(..., conn)

tbl4_prescriptions(..., conn)

tbl4_procedureevents_mv(..., conn)

tbl4_services(..., conn)

tbl4_transfers(..., conn)
}
\arguments{
\item{...}{one or more column names}

\item{conn}{connection}
}
\description{
conn tables

conn tables
}
