% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_tbl_z.R
\name{tbl4}
\alias{tbl4}
\title{Extract table from mimic database}
\usage{
tbl4(tbl, ..., conn)
}
\arguments{
\item{tbl}{one table name or table number}

\item{mimic}{mimic database}

\item{cols}{one or more colnames or column numbers}

\item{schema}{schama name, default is mimiciii}
}
\value{
all table names if tbl is missing, or tible
}
\description{
Extract table from mimic database
}
\examples{
\donttest{
library(mimicR)
connect_MIMIC <- connect_MIMIC()


# extract all table names
tbl4()

# extract services table with row_id and subject_id columns
tbl4('services','row_id','subject_id')

}
}
