% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_segment.R
\name{time_segment}
\alias{time_segment}
\title{time segment for seconds}
\usage{
time_segment(x)
}
\arguments{
\item{x}{seconds}
}
\value{
character
}
\description{
time segment for seconds
}
\examples{
time_segment(10)
time_segment(70)
}
