% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_Unique.R
\name{Unique}
\alias{Unique}
\title{unique rows and numbers
unique rows and numbers for tbl_PostgreSQLConnection}
\usage{
Unique(x, ...)
}
\arguments{
\item{x}{table of tbl_PostgreSQLConnection}

\item{...}{one or more columns, if missing, all column names will be used}
}
\value{
lazy table
}
\description{
unique rows and numbers
unique rows and numbers for tbl_PostgreSQLConnection
}
