% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m3_uploadData_gz.R
\name{uploadData_3gz}
\alias{uploadData_3gz}
\title{upload data to PostgreSQL}
\usage{
uploadData_3gz(
  data_dir,
  tbl = c("admissions", "callout", "caregivers", "chartevents", "cptevents", "d_cpt",
    "d_icd_diagnoses", "d_icd_procedures", "d_items", "d_labitems", "datetimeevents",
    "diagnoses_icd", "drgcodes", "icustays", "inputevents_cv", "inputevents_mv",
    "labevents", "microbiologyevents", "noteevents", "outputevents", "patients",
    "prescriptions", "procedureevents_mv", "procedures_icd", "services", "transfers"),
  conn
)
}
\arguments{
\item{data_dir}{data directory}

\item{tbl}{table names for some one}

\item{conn}{connection, if missing, it will be get 'connectiontopsql' from global environment.}
}
\value{
add 26 data to PostgreSQL
}
\description{
upload data to PostgreSQL
}
\examples{
\donttest{
uploadData_3gz('f:/data')
}
}
