% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_uploadData_gz.R
\name{uploadData_4gz}
\alias{uploadData_4gz}
\title{upload mimic iv data to PostgreSQL}
\usage{
uploadData_4gz(
  data_dir,
  tbl = c("admissions", "chartevents", "d_hcpcs", "d_icd_diagnoses",
    "d_icd_procedures", "d_items", "d_labitems", "datetimeevents", "diagnoses_icd",
    "drgcodes", "emar", "emar_detail", "hcpcsevents", "icustays", "inputevents",
    "labevents", "microbiologyevents", "outputevents", "patients", "pharmacy", "poe",
    "poe_detail", "prescriptions", "procedureevents", "procedures_icd", "services",
    "transfers"),
  conn
)
}
\arguments{
\item{data_dir}{data directory}

\item{tbl}{table names for some one}

\item{conn}{connection, if missing, it will be get 'connectiontopsql' from global environment.}
}
\value{
add mimic iv data to PostgreSQL
}
\description{
upload mimic iv data to PostgreSQL
}
\examples{
\donttest{
uploadData_4gz('f:/data')
}
}
